补件原因代码修改

+++ src/main/java/com/cdkj/loan/ao/ISupplementReasonAO.java
@@ -13,7 +13,7 @@ public interface ISupplementReasonAO {
     static final String DEFAULT_ORDER_COLUMN = "id";
 
     // 新增补件原因
-    public Long addSupplementReason(XN632130Req req);
+    public int addSupplementReason(XN632130Req req);
 
     public Paginable<SupplementReason> querySupplementReasonPage(int start,
             int limit, SupplementReason condition);
@@ -21,7 +21,7 @@ public interface ISupplementReasonAO {
     public List<SupplementReason> querySupplementReasonList(
             SupplementReason condition);
 
-    public SupplementReason getSupplementReason(Long id);
+    public SupplementReason getSupplementReason(int id);
 
     // 历史补件原因
     public List<SupplementReason> querySupplementReason(
